Add node-sha256.
We added the latest node-sha256 version to package.json and then ran
`npm install`.  This added node-sha256.js to the node_modules directory.
We then copied node-sha256.js to the scripts directory in order to make
the library available to our front-end code.

We used npm instead of bower for this dependency because there do not
seem to be any SHA-256 implementations maintained for Bower.
We chose node-sha256 because it is a thin wrapper around Brian Turek's
library, which seems to be popular.  Philip Stark also used this
library in his implementation.